+++ src/components/error-boundary/ErrorBoundary.tsx
@@ -1,19 +1,19 @@
 /** @jsx jsx */
-import { Styled, Flex, Box, jsx } from 'theme-ui'
-import React from 'react'
-import routes from '../../routes'
+import { Styled, Flex, Box, jsx } from 'theme-ui';
+import React from 'react';
+import routes from '../../routes';
 
 class ErrorBoundary extends React.Component {
   state = {
     error: null,
-  }
+  };
 
   static getDerivedStateFromError(error: any) {
-    return { error }
+    return { error };
   }
 
   componentDidCatch(error: any) {
-    console.error('logging', error)
+    console.error('logging', error);
   }
 
   render() {
@@ -22,14 +22,14 @@ class ErrorBoundary extends React.Component {
         <Box>
           <Flex sx={{ pt: 7 }} variant="containers.center">
             <Styled.h1>Something broke</Styled.h1>
-            <Styled.a href={routes.home}>Back to home</Styled.a>
+            <Styled.a href={routes.landing_page}>Back to home</Styled.a>
           </Flex>
         </Box>
-      )
+      );
     }
 
-    return this.props.children
+    return this.props.children;
   }
 }
 
-export default ErrorBoundary
+export default ErrorBoundary;