+++ src/components/menus/Menu.tsx
@@ -1,17 +1,34 @@
 import React, { ReactNode } from 'react';
 import { Box, SxStyleProp } from 'theme-ui';
-import { useDataDispatch, useDataState } from '../../data/DataLayer';
+// import { useDataDispatch, useDataState } from '../../data/DataLayer';
+import { MobileAndTablet, Desktop } from 'react-responsive-simple';
+import { CgMenuGridO } from 'react-icons/cg';
+import { clearAllBodyScrollLocks, disableBodyScroll } from 'body-scroll-lock';
+import { MenuItemType } from '../../pages/landing-page/LandingPage';
 
 export interface MenuProps {
+  sx?: SxStyleProp;
+  schema: Record<'left' | 'right', MenuItemType[]>;
+}
+
+export interface MenuItemProps {
   children: ReactNode;
+  OnHover?: () => JSX.Element;
   sx?: SxStyleProp;
 }
 
-export interface MenuItemProps extends MenuProps {
-  OnHover?: () => JSX.Element;
-}
-
 export type MenuActions = 'setMenuState';
+
+export const MenuItem: React.FC<MenuItemProps> = ({ children, sx }) => {
+  // const state = useDataState();
+  // const dispatch = useDataDispatch();
+
+  return (
+    <Box sx={{ position: 'relative' }}>
+      <Box sx={{ ...sx, variant: 'styles.menuItem' }}>{children}</Box>
+    </Box>
+  );
+};
 
 export const menuReducer = {
   setMenuState: (open: boolean): Record<string, unknown> => {
@@ -19,30 +36,115 @@ export const menuReducer = {
   },
 };
 
-export const Menu: React.SFC<MenuProps> = ({ children, sx }) => {
-  return <Box sx={{ ...sx, variant: 'styles.menu' }}>{children}</Box>;
+export const Menu: React.FC<MenuProps> = ({ schema, sx }) => {
+  return (
+    <Box sx={{ width: '100%', '> div': { width: '100%' } }}>
+      <Desktop>
+        <MenuDesktop schema={schema} sx={sx} />
+      </Desktop>
+
+      <MobileAndTablet>
+        <MenuMobile schema={schema} />
+      </MobileAndTablet>
+    </Box>
+  );
 };
 
-export const MenuItem: React.SFC<MenuItemProps> = ({
-  children,
-  sx,
-  OnHover,
-}) => {
-  const state = useDataState();
-  const dispatch = useDataDispatch();
+export const MenuDesktop: React.FC<MenuProps> = ({ schema, sx }) => {
+  const leftItems = schema.left;
+  const rightItems = schema.right;
 
   return (
-    <Box
-      sx={{ position: 'relative' }}
-      onMouseEnter={() =>
-        dispatch && dispatch({ type: 'setMenuState', payload: { open: true } })
-      }
-      onMouseLeave={() =>
-        dispatch && dispatch({ type: 'setMenuState', payload: { open: false } })
-      }
-    >
-      {state?.open && OnHover && <OnHover></OnHover>}
-      <Box sx={{ ...sx, variant: 'styles.menuItem' }}>{children}</Box>
+    <Box sx={{ ...sx, variant: 'styles.menu' }}>
+      <Box
+        data-testid="headerLeft"
+        sx={{ flexGrow: 1, display: 'inherit', ml: 3 }}
+      >
+        {leftItems &&
+          leftItems.map((leftItem, _) => {
+            return (
+              <MenuItem sx={{ ml: 5 }}>
+                {leftItem.text ?? leftItem.icon}
+              </MenuItem>
+            );
+          })}
+      </Box>
+
+      {rightItems &&
+        rightItems.map((rightItem, _) => {
+          return (
+            <MenuItem sx={{ ml: 5 }}>
+              {rightItem.text ?? rightItem.icon}
+            </MenuItem>
+          );
+        })}
     </Box>
+  );
+};
+
+export const MenuMobile: React.FC<MenuProps> = ({ schema }) => {
+  const [open, setOpen] = React.useState(false);
+
+  const items = [...schema.left, ...schema.right];
+
+  React.useEffect(() => {
+    if (open) {
+      disableBodyScroll(document.head);
+    } else {
+      clearAllBodyScrollLocks();
+    }
+  }, [open]);
+
+  const onClick = () => {
+    setOpen(!open);
+  };
+
+  return (
+    <>
+      <Box
+        onClick={onClick}
+        sx={{
+          position: 'relative',
+          textAlign: 'right',
+          cursor: 'pointer',
+          zIndex: 999999,
+        }}
+      >
+        <CgMenuGridO size="30"></CgMenuGridO>
+      </Box>
+
+      {open && (
+        <Box
+          data-testid="menuOverlay"
+          sx={{
+            position: 'fixed',
+            zIndex: 99,
+            width: '100%',
+            backgroundColor: 'white',
+            maxHeight: '100vh',
+            left: 0,
+            right: 0,
+            bottom: 0,
+            top: 0,
+          }}
+        >
+          <Box
+            sx={{
+              variant: 'styles.menu',
+              flexDirection: 'column',
+              alignItems: 'center',
+              justifyContent: 'center',
+              height: '100vh',
+            }}
+          >
+            {items.map((menuItem, _) => {
+              return (
+                <MenuItem sx={{ mt: 3, fontSize: 4 }}>{menuItem.text}</MenuItem>
+              );
+            })}
+          </Box>
+        </Box>
+      )}
+    </>
   );
 };